refactor: update folder structure and UI
